MAINT: generic operators
