--- data-prepper-expression/src/main/java/org/opensearch/dataprepper/expression/InOperator.java
@@ -1,30 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package org.opensearch.dataprepper.expression;
-
-import org.opensearch.dataprepper.expression.antlr.DataPrepperExpressionParser;
-
-import javax.inject.Named;
-import java.util.Set;
-
-import static com.google.common.base.Preconditions.checkArgument;
-
-@Named
-class InOperator implements Operator<Boolean> {
-    @Override
-    public Integer getSymbol() {
-        return DataPrepperExpressionParser.IN_SET;
-    }
-
-    @Override
-    public Boolean evaluate(final Object... args) {
-        checkArgument(args.length == 2, "Operands length needs to be 2.");
-        if (!(args[1] instanceof Set)) {
-            throw new IllegalArgumentException(args[1] + " should be Set");
-        }
-        return ((Set<?>) args[1]).contains(args[0]);
-    }
-}
